Make custom resource registration function generic
